Newly generated bindings from XML created by Doxygen 1.6.1.
There are almost no changes in this commit except for the dates in the
file headers and minor differences in enums.lisp and typedefs.lisp.
This was a test to see what would happen if I updated Doxygen.  I'm
committing this since we'll keep on using these files and so the diffs
of future commits will be cleaner.